Make ObjectAsDescriptionList its own component
